+++ CHANGELOG.md
@@ -6,13 +6,19 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#122](https://github.com/zfcampus/zf-oauth2/pull/120) adds support for token
+- [#141](https://github.com/zfcampus/zf-oauth2/pull/141) and
+  [#148](https://github.com/zfcampus/zf-oauth2/pull/148) update the component to
+  allow usage with v3 releases of Zend Framework components on which it depends,
+  while maintaining backwards compatibility with v2 components.
+- [#141](https://github.com/zfcampus/zf-oauth2/pull/141) and
+  [#148](https://github.com/zfcampus/zf-oauth2/pull/148) add support for PHP 7.
+- [#122](https://github.com/zfcampus/zf-oauth2/pull/122) adds support for token
   revocation via the `/oauth/revoke` path. The path expects a POST request as
   either urlencoded or JSON values with the parameters:
   - `token`, the access token to revoke
   - `token_type_hint => access_token` to indicate an access token is being
     revoked.
-- [#146](https://github.com/zfcampus/zf-oauth2/pull/120) updates the
+- [#146](https://github.com/zfcampus/zf-oauth2/pull/146) updates the
   `AuthController` to catch `ZF\ApiProblem\Exception\ProblemExceptionInterface`
   instances thrown by the OAuth2 server and return `ApiProblemResponse`s.
 
@@ -22,7 +28,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- Nothing.
+- [#141](https://github.com/zfcampus/zf-oauth2/pull/141) removes support for PHP 5.5.
 
 ### Fixed